pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.2",
    "psycopg2-binary>=2.9.10",
    "pybase64>=1.4.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "supabase>=2.22.0",
    "uvicorn>=0.37.0",
]
